refactor(discord): decompose tmux_watcher.rs into child modules (#3832)
Move the inline `#[cfg(test)] mod tests` body verbatim into a child module
`tmux_watcher/tests.rs`, shrinking the #3565 hot file from 10268 -> 7106 raw
lines. Pure relocation, zero production behavior change, no visibility widening
(the relocated module keeps `use super::*` access to the root coordinator).

This is the only verbatim-movable bulk in the file: the issue's three
production "split boundaries" (terminal-delivery/session-bound glue,
placeholder/status-panel cleanup, monitor-auto-turn) are inline, non-contiguous
code interleaved across the 6863-line `tmux_output_watcher_with_restore`
coordinator and cannot be moved verbatim without behavior-bearing extraction,
so they are intentionally left in place per the understand-phase analysis.

Same blessed test-relocation pattern as #3715 (`tui_prompt_relay/tests.rs`).
- hotfile_ratchet.toml: 10268 -> 7106 (lock in the raw-line win)
- giant baseline: 7103 -> 7106 (+3, the brace-less `#[cfg(test)] #[path] mod
  tests;` include-stub counts as production; documented admission)
