+++ ARCHITECTURE.md
@@ -603,6 +603,7 @@ src/
 │   │   │   ├── terminal_readiness.rs
 │   │   │   ├── terminal_readiness_tests.rs
 │   │   │   ├── terminal_send.rs
+│   │   │   ├── tests.rs
 │   │   │   ├── turn_identity.rs
 │   │   │   ├── turn_identity_tests.rs
 │   │   │   ├── utf8_chunk_decoder.rs
